requirement.txt
requests
mysql-connector-python
pytest